Added user definitions

+++ service/defintitions/users.js
@@ -1,11 +1,60 @@
+const Joi = require('joi');
 const db = require('../config/db');
 
 const User = db.models.user;
 
+async function createUser(call, callback) {
+  const { fullName, email, password } = call.request;
+
+  try {
+    const schema = Joi.object().keys({
+      fullName: Joi.string().optional(),
+      email: Joi.string()
+        .email()
+        .optional(),
+      password: Joi.string().optional(),
+    });
+
+    const validation = Joi.validate({ fullName, email, password }, schema);
+    if (validation.error !== null) throw new Error(validation.error.details[0].message);
+
+    const data = await User.create({ fullName, email, password });
+    callback(null, {
+      success: true,
+      data: JSON.stringify(data, (key, val) => (val !== null ? val : undefined)),
+    });
+  } catch (err) {
+    console.error(err);
+    callback(err);
+  }
+}
+
+async function getUserByID(call, callback) {
+  const { id } = call.request;
+  try {
+    const schema = Joi.object().keys({ id: Joi.string().required() });
+    const validation = Joi.validate({ id }, schema);
+    if (validation.error !== null) throw new Error(validation.error.details[0].message);
+
+    const user = await User.findById(id);
+
+    callback(null, {
+      success: true,
+      data: JSON.stringify(user, (key, val) => (val !== null ? val : undefined)),
+    });
+  } catch (error) {
+    console.error(error);
+    callback(error);
+  }
+}
+
 async function getAllUsers(call, callback) {
   try {
     const data = await User.find();
-    callback(null, { success: true, data: JSON.stringify(data) });
+    callback(null, {
+      success: true,
+      data: JSON.stringify(data, (key, val) => (val !== null ? val : undefined)),
+    });
   } catch (error) {
     console.error(error);
     callback(error);
@@ -17,10 +66,18 @@ async function updateUser(call, callback) {
 
   try {
     const upd = JSON.parse(updates);
+    const schema = Joi.object().keys({
+      id: Joi.string().required(),
+      fullName: Joi.string().optional(),
+      email: Joi.string()
+        .email()
+        .optional(),
+      password: Joi.string().optional(),
+    });
 
-    // TODO: Validate updates...
+    const validation = Joi.validate({ id, ...upd }, schema);
+    if (validation.error !== null) throw new Error(validation.error.details[0].message);
 
-    // Get the user
     const user = await User.findById(id);
 
     // Add all updates to user...
@@ -30,33 +87,13 @@ async function updateUser(call, callback) {
 
     // Save user
     const data = await user.save();
-    callback(null, { success: true, data: JSON.stringify(data) });
-  } catch (err) {
-    console.error(err);
-    callback(err);
-  }
-}
-
-async function createUser(call, callback) {
-  const {
-    firstName, lastName, gender, email, phone, password,
-  } = call.request;
-
-  // Validate user details...
-
-  try {
-    const data = await User.create({
-      firstName,
-      lastName,
-      gender,
-      email,
-      phone,
-      password,
+    callback(null, {
+      success: true,
+      data: JSON.stringify(data, (key, val) => (val !== null ? val : undefined)),
     });
-    callback(null, { success: true, data: JSON.stringify(data) });
-  } catch (err) {
-    console.error(err);
-    callback(err);
+  } catch (error) {
+    console.error(error);
+    callback(error);
   }
 }
 
@@ -66,28 +103,18 @@ async function deleteUserByID(call, callback) {
   // Validate details...
 
   try {
+    const schema = Joi.object().keys({ id: Joi.string().required() });
+    const validation = Joi.validate({ id }, schema);
+    if (validation.error !== null) throw new Error(validation.error.details[0].message);
+
     const data = await User.findByIdAndRemove(id);
-    callback(null, { success: true, data: JSON.stringify(data) });
+    callback(null, {
+      success: true,
+      data: JSON.stringify(data, (key, val) => (val !== null ? val : undefined)),
+    });
   } catch (err) {
     console.error(err);
     callback(err);
-  }
-}
-
-async function getUserByID(call, callback) {
-  const { id } = call.request;
-  try {
-    const user = await User.findById(id);
-
-    if (!user) {
-      callback(null, { success: true, data: null });
-      return;
-    }
-
-    callback(null, { success: true, data: JSON.stringify(user) });
-  } catch (error) {
-    console.error(error);
-    callback(error);
   }
 }
